+++ cmd/sense/main.go
@@ -3,17 +3,18 @@ package main
 import (
 	"context"
 	"fmt"
+	"os"
 
 	"github.com/gregoryv/filechange"
 )
 
 func main() {
-	s := new(filechange.Sensor)
-	s.Visit = func(modified ...string) {
+	root, _ := os.Getwd()
+	s := filechange.NewSensor(root, func(modified ...string) {
 		for _, f := range modified {
 			fmt.Println(f)
 		}
-	}
+	})
 	s.Recursive = true
 	s.Run(context.Background())
 }