+++ example_test.go
@@ -3,19 +3,16 @@ package filechange_test
 import (
 	"context"
 	"fmt"
-	"time"
 
 	"github.com/gregoryv/filechange"
 )
 
 func Example() {
-	s := &filechange.Sensor{
-		Visit: func(modified ...string) {
-			fmt.Println(modified)
-		},
-		Recursive: true,
-		Pause:     2 * time.Second,
-	}
+	s := filechange.NewSensor(".", func(modified ...string) {
+		fmt.Println(modified)
+	})
+	s.Recursive = true
+
 	ctx, _ := context.WithCancel(context.Background())
 	go s.Run(ctx)
 }